Fixed performance from custom streams

+++ java/clearley/CharParseStream.java
@@ -1,9 +1,9 @@
 package clearley;
 
 /**
- * A parse stream that returns chars, except it actually returns Longs because Clojure doesn't support
+ * A parse stream that returns chars, except it actually returns ints because Clojure doesn't support
  * fast primitive char casts!
  */
 public interface CharParseStream extends ParseStream {
-	long current();
+	int current();
 }
